fix(plugins): label plugin failure handling honestly
Move runtime plugin exception handling behind an error-containment module while keeping the legacy sandbox import as a compatibility alias.

Constraint: #1597 requires sandbox labeling to avoid implying process or VM isolation.\nRejected: Removing sandbox.ts outright | existing imports and third-party callers may still rely on the legacy path.\nConfidence: high\nScope-risk: narrow\nDirective: Do not describe plugin error containment as a security sandbox unless a real isolation boundary is added.\nTested: bunx tsc --noEmit; bun test tests/http/health tests/http/logging tests/plugins/sandbox.test.ts tests/plugins/path-containment.test.ts\nNot-tested: External plugins importing the deprecated sandbox path.
