+++ tests/plugins/sandbox.test.ts
@@ -1,8 +1,9 @@
 import { describe, expect, test } from "bun:test";
 import { PluginEventBus, type PluginEventMap } from "../../src/plugins/event-bus.ts";
+import { runPluginWithErrorContainment } from "../../src/plugins/error-containment.ts";
 import { runPluginSandbox } from "../../src/plugins/sandbox.ts";
 
-describe("runPluginSandbox", () => {
+describe("plugin error containment", () => {
   test("returns plugin failures and emits plugin:error without throwing", async () => {
     const bus = new PluginEventBus();
     const events: PluginEventMap["plugin:error"][] = [];
@@ -10,8 +11,8 @@ describe("runPluginSandbox", () => {
       events.push(event);
     });
 
-    const success = await runPluginSandbox({ plugin: "ok-plugin", phase: "runtime", eventBus: bus }, () => "ok");
-    const failure = await runPluginSandbox({ plugin: "bad-plugin", phase: "init", eventBus: bus }, () => {
+    const success = await runPluginWithErrorContainment({ plugin: "ok-plugin", phase: "runtime", eventBus: bus }, () => "ok");
+    const failure = await runPluginWithErrorContainment({ plugin: "bad-plugin", phase: "init", eventBus: bus }, () => {
       throw new Error("boom");
     });
     const throwingBus = {
@@ -19,7 +20,7 @@ describe("runPluginSandbox", () => {
         throw new Error("observer failed");
       },
     } as Pick<PluginEventBus, "emit">;
-    const observerFailure = await runPluginSandbox({ plugin: "observer-plugin", phase: "destroy", eventBus: throwingBus }, () => {
+    const observerFailure = await runPluginWithErrorContainment({ plugin: "observer-plugin", phase: "destroy", eventBus: throwingBus }, () => {
       throw "plain failure";
     });
 
@@ -29,5 +30,10 @@ describe("runPluginSandbox", () => {
     expect(events).toHaveLength(1);
     expect(events[0]).toMatchObject({ plugin: "bad-plugin", phase: "init", message: "boom" });
     expect(events[0].error).toBeInstanceOf(Error);
+  });
+
+  test("keeps the legacy sandbox export as a compatibility alias", async () => {
+    await expect(runPluginSandbox({ plugin: "legacy", phase: "runtime" }, () => "ok"))
+      .resolves.toEqual({ ok: true, value: "ok" });
   });
 });
